requirements.txt
streamlit
google-generativeai
sounddevice
speechrecognition
numpy
pyttsx3
requests
python-dotenv
wave
